# src/main/resources/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: 콘서트 예약 서비스 API
  version: 1.0.0
paths:
  /tokens:
    post:
      summary: 토큰 발급 API
      operationId: enrollToken
      parameters:
        - in: header
          name: X-USER-TOKEN
          schema:
            type: string
            minLength: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenHttpRequest'
      responses:
        '200':
          description: 새로 발급되거나 초기화된 사용자 인증 토큰 반환
          headers:
            tokenId:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenHttpResponse'

  /concerts/{concertId}/queues:
    get:
      summary: 대기열 확인 API
      operationId: getConcertQueuePosition
      parameters:
        - in: header
          name: X-QUEUE-TOKEN
          required: true
          schema:
            type: string
            minLength: 1
        - in: path
          name: concertId
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: 대기열 남은 순서 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueuePositionHttpResponse'

    post:
      summary: 대기열 추가 API
      operationId: enrollConcertQueue
      parameters:
        - in: header
          name: X-USER-TOKEN
          required: true
          schema:
            type: string
            minLength: 1
        - in: path
          name: concertId
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: 좌석 임시 배정 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueEnrollmentHttpResponse'

  /concerts/{concertId}/available-dates:
    get:
      summary: 예약 가능 날짜 조회 API
      operationId: selectConcertAvailableDates
      parameters:
        - in: header
          name: X-QUEUE-TOKEN
          required: true
          schema:
            type: string
            minLength: 1
        - in: path
          name: concertId
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: 예약 가능한 날짜 목록 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailableDatesHttpResponse'

  /concerts/{concertId}/details/{detailId}/available-seats:
    get:
      summary: 예약 가능 좌석 조회 API
      operationId: selectConcertAvailableSeats
      parameters:
        - in: header
          name: X-QUEUE-TOKEN
          required: true
          schema:
            type: string
            minLength: 1
        - in: path
          name: concertId
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
        - in: path
          name: detailId
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: 예약 가능한 좌석 목록 반환
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AvailableSeatsHttpResponse'

  /concerts/{concertId}/details/{detailId}/seats/{seatId}:
    post:
      summary: 좌석 예약 API
      operationId: reserveConcertSeat
      parameters:
        - in: header
          name: X-QUEUE-TOKEN
          required: true
          schema:
            type: string
            minLength: 1
        - in: path
          name: concertId
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
        - in: path
          name: detailId
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
        - in: path
          name: seatId
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: 좌석 임시 배정 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatReservationHttpResponse'

  /points:
    post:
      summary: 잔액 충전 API
      operationId: chargePoint
      parameters:
        - in: header
          name: X-USER-TOKEN
          required: true
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PointChargeHttpRequest'
      responses:
        '200':
          description: 충전 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PointChargeHttpResponse'

    get:
      summary: 잔액 조회 API
      operationId: selectPoint
      parameters:
        - in: header
          name: X-USER-TOKEN
          required: true
          schema:
            type: string
            minLength: 1
        - in: query
          name: userId
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: 현재 잔액 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PointSelectHttpResponse'

  /orders/payments:
    post:
      summary: 결제 API
      operationId: executePayment
      parameters:
        - in: header
          name: X-QUEUE-TOKEN
          required: true
          schema:
            type: string
            minLength: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentHttpRequest'
      responses:
        '200':
          description: 결제 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentHttpResponse'

components:
  schemas:
    TokenHttpRequest:
      type: object
      required:
        - userId
      properties:
        userId:
          type: integer
          example: 1
          format: int64
          minimum: 1
    TokenHttpResponse:
      type: object
      properties:
        token:
          type: string
          example: "21113572-4cb4-44ef-8e19-63269724b8be"

    QueuePositionHttpResponse:
      type: object
      properties:
        count:
          type: integer
          example: 5

    QueueEnrollmentHttpResponse:
      type: object
      properties:
        message:
          type: string
          example: "대기열에 등록되었습니다."

    AvailableDatesHttpResponse:
      type: array
      items:
        type: string
        example: "2024-10-20"

    AvailableSeatsHttpResponse:
      type: object
      properties:
        seatId:
          type: integer
          example: 3
          format: int64
        seatName:
          type: string
          example: "A-3"
        price:
          type: integer
          example: 15000

    SeatReservationHttpResponse:
      type: object
      properties:
        message:
          type: string
          example: "좌석이 임시 배정되었습니다."

    PointChargeHttpRequest:
      type: object
      required:
        - userId
        - amount
      properties:
        userId:
          type: integer
          example: 1
          format: int64
          minimum: 1
        amount:
          type: integer
          example: 20000
          minimum: 1
    PointChargeHttpResponse:
      type: object
      properties:
        message:
          type: string
          example: "포인트 충전이 완료되었습니다."

    PointSelectHttpResponse:
      type: object
      properties:
        point:
          type: integer
          example: 25000

    PaymentHttpRequest:
      type: object
      required:
        - orderId
      properties:
        orderId:
          type: integer
          example: 1
          format: int64
          minimum: 1
    PaymentHttpResponse:
      type: object
      properties:
        message:
          type: string
          example: 결제가 완료되었습니다.